+++ csp/dummy/deploy/clusterrole.yaml
@@ -52,4 +52,4 @@ rules:
 # the plugin has finished deleting them.
 - apiGroups: [""]
   resources: ["namespaces"]
-  verbs: ["get", "list", "watch", "create", "update", "patch", "delete"]
+  verbs: ["get", "list", "watch", "delete"]
