+++ gateway/config/k8s-dev-setup/regional/00-namespace.yaml
@@ -4,7 +4,7 @@
 # The gateway provisions this namespace on the first tenant-scoped write, so it is NOT
 # hand-managed. It is declared here only so the storage SKU and image catalogs can be applied
 # before the gateway is up. The owner label must be secapi.cloud/tenant — the key
-# namespaceOwnedBy checks; any other key makes the namespace un-adoptable.
+# namespaceOwnedBy checks; with any other key the controller treats it as foreign.
 apiVersion: v1
 kind: Namespace
 metadata:
