+++ resource/network/v1/network/backend/kubernetes/controller.go
@@ -1,6 +1,7 @@
 package kubernetes
 
 import (
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -9,7 +10,18 @@ import (
 	builder "github.com/eu-sovereign-cloud/ecp/framework/backend/kubernetes/builder"
 	frameworkcontroller "github.com/eu-sovereign-cloud/ecp/framework/backend/kubernetes/controller"
 	netdom "github.com/eu-sovereign-cloud/ecp/resource/network/v1/network"
+	routetablek8s "github.com/eu-sovereign-cloud/ecp/resource/network/v1/route-table/backend/kubernetes"
+	subnetk8s "github.com/eu-sovereign-cloud/ecp/resource/network/v1/subnet/backend/kubernetes"
 )
+
+// ChildResourceGVRs is the closed set of SECA types that live in the namespace a Network owns for
+// its children. It gates the gateway's refusal to delete a non-empty Network (409) and the
+// controller's re-check before it deletes that namespace, so a type missing from it makes the
+// namespace look empty when it is not — and the namespace delete cascades.
+var ChildResourceGVRs = []schema.GroupVersionResource{
+	routetablek8s.RouteTableGVR,
+	subnetk8s.SubnetGVR,
+}
 
 // Controller drives network reconciliation using the GenericController.
 type Controller struct {
@@ -47,11 +59,6 @@ func NewController(
 			options.MaxConditions,
 		),
 	}
-	c.WithEnsure(k8sadapter.NamespaceEnsure[*netdom.Network](
-		clientset,
-		options.Logger,
-		k8sadapter.NetworkChildren,
-	))
 	c.WithCleanup(k8sadapter.NamespaceCleanup[*netdom.Network](
 		dynClient,
 		clientset,
