+++ resource/workspace/v1/backend/kubernetes/controller.go
@@ -1,6 +1,7 @@
 package kubernetes
 
 import (
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -8,8 +9,36 @@ import (
 	k8sadapter "github.com/eu-sovereign-cloud/ecp/framework/backend/kubernetes"
 	builder "github.com/eu-sovereign-cloud/ecp/framework/backend/kubernetes/builder"
 	frameworkcontroller "github.com/eu-sovereign-cloud/ecp/framework/backend/kubernetes/controller"
+	instancek8s "github.com/eu-sovereign-cloud/ecp/resource/compute/v1/instance/backend/kubernetes"
+	internetgatewayk8s "github.com/eu-sovereign-cloud/ecp/resource/network/v1/internet-gateway/backend/kubernetes"
+	netk8s "github.com/eu-sovereign-cloud/ecp/resource/network/v1/network/backend/kubernetes"
+	nick8s "github.com/eu-sovereign-cloud/ecp/resource/network/v1/nic/backend/kubernetes"
+	publicipk8s "github.com/eu-sovereign-cloud/ecp/resource/network/v1/public-ip/backend/kubernetes"
+	securitygrouprulek8s "github.com/eu-sovereign-cloud/ecp/resource/network/v1/security-group-rule/backend/kubernetes"
+	securitygroupk8s "github.com/eu-sovereign-cloud/ecp/resource/network/v1/security-group/backend/kubernetes"
+	bsk8s "github.com/eu-sovereign-cloud/ecp/resource/storage/v1/block-storage/backend/kubernetes"
+	imgk8s "github.com/eu-sovereign-cloud/ecp/resource/storage/v1/image/backend/kubernetes"
 	wsdom "github.com/eu-sovereign-cloud/ecp/resource/workspace/v1"
 )
+
+// ChildResourceGVRs is the closed set of SECA types that live in the namespace a Workspace owns
+// for its children. It gates the gateway's refusal to delete a non-empty Workspace (409) and the
+// controller's re-check before it deletes that namespace, so a type missing from it makes the
+// namespace look empty when it is not — and the namespace delete cascades.
+//
+// Subnet and RouteTable are absent on purpose: they are network-scoped and live in the Network's
+// own child namespace. See network's ChildResourceGVRs.
+var ChildResourceGVRs = []schema.GroupVersionResource{
+	bsk8s.BlockStorageGVR,
+	imgk8s.ImageGVR,
+	netk8s.NetworkGVR,
+	nick8s.NICGVR,
+	publicipk8s.PublicIPGVR,
+	internetgatewayk8s.InternetGatewayGVR,
+	securitygroupk8s.SecurityGroupGVR,
+	securitygrouprulek8s.SecurityGroupRuleGVR,
+	instancek8s.InstanceGVR,
+}
 
 // Controller drives workspace reconciliation using the GenericController.
 type Controller struct {
@@ -47,11 +76,6 @@ func NewController(
 			options.MaxConditions,
 		),
 	}
-	c.WithEnsure(k8sadapter.NamespaceEnsure[*wsdom.Workspace](
-		clientset,
-		options.Logger,
-		k8sadapter.WorkspaceChildren,
-	))
 	c.WithCleanup(k8sadapter.NamespaceCleanup[*wsdom.Workspace](
 		dynClient,
 		clientset,
