+++ resource/workspace/v1/backend/kubernetes/controller_test.go
@@ -162,37 +162,6 @@ func TestWorkspaceController_Reconcile(t *testing.T) {
 		require.Equal(t, k8srt.Result{RequeueAfter: requeueAfter}, res)
 	})
 
-	t.Run("should run the ensure hook on a live resource", func(t *testing.T) {
-		mc := gomock.NewController(t)
-		defer mc.Finish()
-
-		mockRepo := NewMockRepo[*wsdom.Workspace](mc)
-		mockRepo.EXPECT().UpdateStatus(gomock.Any(), gomock.Any()).Return(nil, nil).Times(1)
-
-		fakeClient := fake.NewClientBuilder().
-			WithScheme(newScheme()).
-			WithObjects(newK8sResource()).
-			Build()
-
-		handler := NewWorkspacePluginHandler(mockRepo, NewMockWorkspacePlugin(mc), 1)
-		gc := frameworkcontroller.NewGenericController[*wsdom.Workspace](
-			fakeClient, WorkspaceFromCR, handler, &Workspace{}, 0,
-			slog.New(slog.NewTextHandler(&bytes.Buffer{}, nil)), 1,
-		)
-
-		calls := 0
-		gc.WithEnsure(func(_ context.Context, ws *wsdom.Workspace) error {
-			calls++
-			require.Equal(t, testName, ws.Name)
-			return nil
-		})
-
-		res, err := gc.Reconcile(t.Context(), req)
-		require.NoError(t, err)
-		require.Equal(t, k8srt.Result{}, res)
-		require.Equal(t, 1, calls)
-	})
-
 	// The cleanup hook is the seam that lets a resource tear down the namespace it owns for its
 	// children. It must run only after the plugin has finished deleting, and only while the
 	// finalizer still holds the object — otherwise the side effect is lost.
@@ -234,10 +203,6 @@ func TestWorkspaceController_Reconcile(t *testing.T) {
 
 		calls := 0
 		gc := newDeletingController(t, fakeClient, 0)
-		gc.WithEnsure(func(context.Context, *wsdom.Workspace) error {
-			require.Fail(t, "ensure must not run while the resource is deleting")
-			return nil
-		})
 		gc.WithCleanup(func(ctx context.Context, ws *wsdom.Workspace) error {
 			calls++
 			require.Equal(t, testName, ws.Name)
